+++ CHANGELOG.md
@@ -1,6 +1,6 @@
-# [NEXT]
+# [2.2.2]
 - Improvements in `DirectionAnimation` mixin.
-- Improvements in `Movement` mixin.
+- Improvements in `Movement` mixin. Adds `onMove` method, you can override this to listen component movements
 
 # [2.2.1]
 - Fix `WithSpriteAnimation`